Post-process refactor #2

+++ Suruga/Audio/Commands/Playback/PitchAudioCommand.cs
@@ -1,3 +1,3 @@
 ﻿namespace Suruga.Audio.Commands.Playback;
 
-internal sealed record PitchAudioCommand(double Value) : PlaybackCommand;
+internal sealed record PitchAudioCommand(float Value) : PlaybackCommand;
